# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: AgroGuard Limited - Agricultural IoT Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List all AI-generated recommendations
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
    post:
      operationId: createRecommendation
      tags: [recommendations]
      summary: Create AI crop/irrigation recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "201":
          description: Recommendation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
  /recommendations/{id}/apply:
    patch:
      operationId: applyRecommendation
      tags: [recommendations]
      summary: Mark recommendation as applied
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendation applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Recommendation not found

  # ── STAFF ────────────────────────────────────────────────────────────────────
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffCreateResult"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          description: Staff not found
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Remove staff member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Staff deleted
        "404":
          description: Staff not found

  # ── DASHBOARD ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Recommendation:
      type: object
      required: [id, farmerId, category, title, description, status, createdAt]
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        deviceId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [irrigation, crop, pest, disease, climate, fertilizer]
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, applied, dismissed]
        appliedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    RecommendationInput:
      type: object
      required: [farmerId, category, title, description]
      properties:
        farmerId:
          type: integer
        deviceId:
          type: integer
        category:
          type: string
          enum: [irrigation, crop, pest, disease, climate, fertilizer]
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [low, medium, high]

    # ── STAFF ────────────────────────────────────────────────────────────────────
    StaffMember:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, admin, agronomist, staff]
        status:
          type: string
          enum: [active, inactive]
        department:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StaffCreateResult:
      type: object
      required: [id, name, email, role, status, createdAt, tempPassword]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, admin, agronomist, staff]
        status:
          type: string
          enum: [active, inactive]
        department:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        tempPassword:
          type: string
          description: One-time temporary password shown only at creation; the staff member must change it on first login.
    StaffInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [super_admin, admin, agronomist, staff]
        department:
          type: string
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [super_admin, admin, agronomist, staff]
        status:
          type: string
          enum: [active, inactive]
        department:
          type: string

    # ── DASHBOARD ────────────────────────────────────────────────────────────────
...