# lib/api-zod/src/generated/types/staffCreateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgroGuard Limited - Agricultural IoT Platform API
 * OpenAPI spec version: 0.2.0
 */
import type { StaffCreateResultRole } from './staffCreateResultRole';
import type { StaffCreateResultStatus } from './staffCreateResultStatus';

export interface StaffCreateResult {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: StaffCreateResultRole;
  status: StaffCreateResultStatus;
  /** @nullable */
  department?: string | null;
  createdAt: Date;
  /** One-time temporary password shown only at creation; the staff member must change it on first login. */
  tempPassword: string;
}